// src/app/features/patient/models/patient.model.ts
export interface PatientModel {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  age: number;
  gender: string;
  bloodGroup: string;
  address: string;
  assignedDoctorId: string;
  assignedDoctorName: string;
  departmentId: string;
  departmentName: string;
  status: string;
  createdAt: string;
}